requirements.txt
xgboost
fastapi
uvicorn
scikit-learn
pandas
numpy
python-dotenv
email-validator
matplotlib
streamlit
